refactor(api-compare): move the canonical baseline serializer into its own module
Keeps body-pins from importing a lint gate for a formatting helper, and
adds a repo-level test asserting every committed baseline is canonical.
